pyproject.toml
[project]
name = "restbase"
version = "0.1.0"
description = "Sync any REST API to PostgreSQL with schema inference"
authors = [
  { name="Soumyadeep Das", email="soumyadeep.das423@gmail.com" }
]
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.7"
dependencies = [
    "requests",
    "psycopg2-binary",
    "pyyaml"
]

[project.scripts]
restbase = "restbase.cli:main"

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"
